src/backend/Archaios.AI.LiDARProcessor/requirements.txt
# Core dependencies
numpy==1.26.4
azure-storage-blob>=12.13.0
azure-storage-queue>=12.4.0
azure-identity>=1.10.0
python-dotenv>=0.21.0
laspy>=2.4.3
lazrs>=0.5.1
aiohttp>=3.8.2

# For E57 conversion
pye57>=0.3.0
scipy>=1.8.0

# For geospatial processing
rasterio>=1.3.0
pyproj>=3.3.0
utm>=0.7.0
matplotlib>=3.5.0
Pillow>=9.0.0
scikit-learn>=1.1.0

# For terrain-aware filtering
scikit-image>=0.20.0
